Issue #167: Really make sure the last value is always present
Even with previous attempts to fix this issue, it still persisted for me
(though I didn't notice it on all captures, but it could be that I
didn't notice this if there was a captured value near the end). I
suspect that perhaps the RLE decoder is unaffected or something, my Bus
Pirate uses unencoded output.

What happened was that the duplicate sample removal in CapturedData
would kill the last sample, because it is (by the way EqualityFilter
works) always identical to the next to last sample. However, it would
still record its timestamp for the purposes of the "do we need to add a
final sample" check, so no final sample would be added. This would break
the signal display, truncating it at the last change in a signal.